Refactor management API impl
